refactor: ConsentsData > ConsentsDataPage
